refactor core to use index.ts
